Move puzzle building tests to hyperpuzzle_library
